requirements.txt
django
django-rest-framework
jinja2